django/requirements.txt
asgiref==3.6.0
bibtexparser==1.4.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
django-model-utils==4.3.1
feedparser==6.0.10
gunicorn==20.1.0
idna==3.4
pyparsing==3.0.9
pytz==2023.3
pyzotero==1.5.9
requests==2.30.0
sgmllib3k==1.0.0
sqlparse==0.4.4
urllib3==2.0.2
